Debugging the two-way coupling between SWAN-Thetis
The following errors were detected from utilising the '-g -O0' flags when compiling SWAN and bmi-swan and the "PYTHONMALLOC=malloc mpiexec -np $cores python coupled_model_01.py"
1)The allocatable array XYTST, indicating the number of Test points, was not allocated in swanpre1.ftn L:1865
Solution: We allocated zero Test points
IF (.NOT.ALLOCATED(XYTST)) ALLOCATE(XYTST(0))
Potential problems: We haven't deallocated it, which may cause problems if the command TEST is included in our SWAN configuration file

2) When we read the hotfile with INIT HOTS SING when we run a parallel SWAN, we had a problem in swanpre1.ftn with KGRPNT being out of index. We identified the following undeclared variables
Integer JX, JY, IY, IX, IINDX
However, the problem persisted.
Solution: Utilise INIT HOTS MULT

3)The allocatable array 'FRIC' is not allocated (swanmain.ftn, L.1030)
Solution: We allocate it
IF (.NOT.ALLOCATED(FRIC)) ALLOCATE(FRIC(MXG(4)*MYG(4)))

4) The allocatable array 'NPLAF' is not allocated (swanmain.ftn, L.1032)
Solution: We allocate it
IF (.NOT.ALLOCATED(NPLAF)) ALLOCATE(NPLAF(MXG(11)*MYG(11)))

5) The allocatable array 'TURBF' is not allocated (swanmain.ftn, L.1033)
Solution: We allocate it
IF (.NOT.ALLOCATED(TURBF)) ALLOCATE(TURBF(MXG(12)*MYG(12)))

6) The allocatable array 'MUDLF) is not allocated (swanmain.ftn, L.1034)
Solution: We allocate it
IF (.NOT.ALLOCATED(MUDLF)) ALLOCATE(MUDLF(MXG(13)*MYG(13)))

7) The allocatable array 'AICEF' is not allocated (swanmain.ftn, L.1036)
Solution: We allocate it
IF (.NOT.ALLOCATED(AICEF)) ALLOCATE(AICEF(MXG(14)*MYG(14)))

8) The allocatable array 'HICEF' is not allocated (swanmain.ftn, L.1037)
Solution: We allocate it
IF (.NOT.ALLOCATED(HICEF)) ALLOCATE(HICEF(MXG(15)*MYG(15)))

9)SwanBmiComm.f90, L.1078: Index '-99' of dimension 1 of array 'THQB' below lower bound of 1
Explanation: We have assigned the value -99 in the points that are outside of the SWAN domain. However, THINDX array basically is a pointer array to re-order the SWAN points per rank in the global array in the correct order. However, if THIND(IP) = -99, then the point does not have a value and we shouldn't do anything (we have assigned the value 0 to all element of the THQB array)
Solution:
               IF (THINDX(IP).NE. -99) THEN
!                  THQB(THINDX(IP))=0.0
!                  THHS(THINDX(IP))=0.0
!                  THDIR(THINDX(IP))=0.0
!                  THWLEN(THINDX(IP))=0.0
!               ELSE
                  THQB(THINDX(IP))=THQBMST(IP)
                  THHS(THINDX(IP))=THHSMST(IP)
                  THDIR(THINDX(IP))=THDIRMST(IP)
                  THWLEN(THINDX(IP))=THWLENMST(IP)
               ENDIF